돌 던지고 받기
돌의 조건 추가
돌이 던져진 다음 다시 받는 것까지 완성
