+++ Assets/01.Scripts/Stone.cs.meta
